Revert "VZ-5864: skip deleted pods for Rancher upgrade work around (#3228)"
This reverts commit 685994927c13f038bc659ae36494bb33083ff01b.

+++ platform-operator/controllers/verrazzano/component/rancher/rancher.go
@@ -152,14 +152,9 @@ func checkRancherUpgradeFailure(c client.Client, log vzlog.VerrazzanoLogger) err
 
 	// Check the logs of each pod
 	for i, pod := range podList.Items {
-		// Skip pods that are already being deleted
-		if pod.DeletionTimestamp != nil {
-			continue
-		}
-
 		// Skip pods that are not ready, they will get checked again in another call to isReady.
 		if !isPodReady(pod) {
-			continue
+			break
 		}
 
 		// Get the pod log stream
@@ -173,9 +168,7 @@ func checkRancherUpgradeFailure(c client.Client, log vzlog.VerrazzanoLogger) err
 		restartPod := false
 		scanner := bufio.NewScanner(logStream)
 		for scanner.Scan() {
-			token := scanner.Text()
-			if strings.Contains(token, "Failed to find system chart") {
-				log.Infof("Rancher IsReady: Failed to find system chart for pod %s: %s", pod.Name, token)
+			if strings.Contains(scanner.Text(), "Failed to find system chart") {
 				restartPod = true
 				break
 			}
@@ -243,7 +236,11 @@ func deleteClusterRepos(log vzlog.VerrazzanoLogger) error {
 		return err
 	}
 
+	// Only need to delete the custom resources if the default branch is still release-2.5
 	log.Infof("Rancher IsReady: The default release branch is currently set to %s", defaultBranch)
+	if defaultBranch != "release-v2.5" {
+		return nil
+	}
 
 	// Delete settings.management.cattle.io chart-default-branch
 	err = dynamicClient.Resource(gvr).Delete(context.TODO(), name, metav1.DeleteOptions{})
